pyproject.toml
[project]
name = "bumblebee"
version = "0.1.0"
description = "The first entitative harness — Gemma 4-native persistent digital entities"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "apscheduler>=3.10",
    "aiosqlite>=0.20",
    "numpy>=1.26",
    "click>=8.1",
    "discord-py>=2.3",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
bumblebee = "bumblebee.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bumblebee"]

[tool.hatch.build.targets.sdist]
include = [
    "bumblebee/**",
    "configs/**",
    "docs/**",
    "tests/**",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
